move function to extract.py
